extract servant code from fuzz machinery
